requirements.txt
googlemaps==2.5.1
nltk==3.4.5
pandas==0.25.1
mysql-connector-python==8.0.18
numpy==1.16.5
scikit-learn==0.21.3
flake8==3.7.9
pytest==5.4.1
python-dotenv==0.12.0
black==19.10b0
slackclient==2.5.0
xgboost==1.0.0
Flask==1.1.2
pymessenger==0.0.7.0

